requirements.txt
.# FastAPI Platform Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
PyJWT>=2.10.0
bcrypt>=4.0.0

# Database
motor>=3.7.0
pymongo>=4.14.0

# Task Queue and Scheduling
celery[redis]>=5.3.0
redis>=4.5.0

# Existing website analysis dependencies
aiohttp>=3.8.0
beautifulsoup4>=4.11.0
pydantic>=1.10.0
pydantic-settings>=2.0.0
tqdm>=4.64.0
openai>=1.0.0
python-dotenv>=1.0.0

# Additional utilities
dnspython>=2.8.0
brotli>=1.1.0
tiktoken>=0.5.0
markdownify>=0.11.0